+++ channels/views_test.py
@@ -11,7 +11,9 @@ pytestmark = pytest.mark.betamax
 def test_api_exception(client, logged_in_profile, mocker):
     """Make sure APIExceptions which aren't recognized become 500 errors"""
     exception = APIException("bizarre", "A bizarre exception", "bizarre_field")
-    mocker.patch("channels.serializers.CommentSerializer.update", side_effect=exception)
+    mocker.patch(
+        "channels.serializers.comments.CommentSerializer.update", side_effect=exception
+    )
     url = reverse("comment-detail", kwargs={"comment_id": "e8h"})
     with pytest.raises(APIException) as ex:
         client.patch(url, type="json", data={"text": "updated text"})
